+++ Milky/src/Milky/Scene/Scene.h
@@ -28,7 +28,7 @@ namespace Milky {
 		template<typename Comp>
 		void OnComponentAdded(Entity entity, Comp& component);
 
-		int GetNumEntites();
+		size_t GetNumEntites();
 	private:
 		entt::registry m_Registry;
 		uint32_t m_ViewportWidth = 0, m_ViewportHeight = 0;
